requirements/development.txt
# Development only
pyinotify
git-pylint-commit-hook
django-debug-toolbar
ipdb

# Development and Production
Django
psycopg2
django-mptt
-e git+https://github.com/areski/django-pagination.git@fix-xrange-slice#egg=linaro-django-pagination
webstack-django-sorting
djangorestframework
Pillow
-e git+https://github.com/helber/django-dbsettings.git@master#egg=django-dbsettings
django-modeltranslation
sorl-thumbnail
jsonfield
-e git+https://github.com/awecode/njango.git@master#egg=njango